feat: export convertLatexToMarkup as a standalone bundle
The `convertLatexToMarkup` function is server compatible but is exported
in the main bundle `mathlive.*.*js`. This bundle also imports from
`public/*` which has global code accessing `document`.

Therefore, just by importing the main bundle on server creates an issue.

This commit adds a separate `dist/convert-latex-to-markup.*js` bundle
which can be used on the server.